requirements.txt
playwright==1.47.0
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1
rich==13.9.4

# dev / testing
pytest>=8.0
pytest-asyncio>=0.23
